+++ elenasrc2/engine/x86jitcompiler.h
@@ -181,7 +181,7 @@ public:
 
    virtual void loadNativeCode(_BinaryHelper& helper, MemoryWriter& writer, _Module* binary, _Memory* section);
 
-   virtual void setStaticRootCounter(_JITLoader* loader, int counter, bool virtualMode);
+   virtual void setStaticRootCounter(_JITLoader* loader, size_t counter, bool virtualMode);
 
    x86JITCompiler(bool debugMode);
 };
